Add a check for service numbers array being empty

+++ module/Api/src/Domain/CommandHandler/Bus/UpdateServiceDetails.php
@@ -11,7 +11,7 @@ use Dvsa\Olcs\Transfer\Command\CommandInterface;
 use Doctrine\ORM\Query;
 use Doctrine\Common\Collections\ArrayCollection;
 use Dvsa\Olcs\Api\Entity\Bus\BusReg;
-use Dvsa\Olcs\Api\Entity\Bus\BusNoticePeriod;
+use Dvsa\Olcs\Api\Entity\Bus\BusNoticePeriod as BusNoticePeriodEntity;
 use Dvsa\Olcs\Api\Entity\Bus\BusRegOtherService;
 use Dvsa\Olcs\Api\Entity\Bus\BusServiceType as BusServiceTypeEntity;
 use Dvsa\Olcs\Transfer\Command\Bus\UpdateServiceDetails as UpdateServiceDetailsCmd;
@@ -82,7 +82,7 @@ final class UpdateServiceDetails extends AbstractCommandHandler implements Trans
             $command->getReceivedDate(),
             $command->getEffectiveDate(),
             $command->getEndDate(),
-            $busNoticePeriod,
+            $this->getRepo()->getReference(BusNoticePeriodEntity::class, $busNoticePeriod),
             $busRules
         );
 
@@ -129,36 +129,38 @@ final class UpdateServiceDetails extends AbstractCommandHandler implements Trans
     {
         $reduced = [];
 
-        foreach ($otherServiceNumbers as $serviceNumber) {
-            if (empty($serviceNumber['serviceNo'])) {
-                // filter out empty values
-                continue;
+        if (!empty($otherServiceNumbers)) {
+            foreach ($otherServiceNumbers as $serviceNumber) {
+                if (empty($serviceNumber['serviceNo'])) {
+                    // filter out empty values
+                    continue;
+                }
+
+                if (!empty($serviceNumber['id'])) {
+                    // update
+                    /** @var BusRegOtherService $otherServiceEntity */
+                    $otherServiceEntity = $this->getRepo('BusRegOtherService')->fetchById(
+                        $serviceNumber['id'],
+                        Query::HYDRATE_OBJECT,
+                        $serviceNumber['version']
+                    );
+                    $otherServiceEntity->setServiceNo($serviceNumber['serviceNo']);
+                } else {
+                    // create
+                    $otherServiceEntity = new BusRegOtherService();
+                    $otherServiceEntity->setBusReg($busReg);
+                    $otherServiceEntity->setServiceNo($serviceNumber['serviceNo']);
+                }
+
+                $this->getRepo('BusRegOtherService')->save($otherServiceEntity);
+                $reduced[] = $otherServiceEntity->getId();
             }
 
-            if (!empty($serviceNumber['id'])) {
-                // update
-                /** @var BusRegOtherService $otherServiceEntity */
-                $otherServiceEntity = $this->getRepo('BusRegOtherService')->fetchById(
-                    $serviceNumber['id'],
-                    Query::HYDRATE_OBJECT,
-                    $serviceNumber['version']
-                );
-                $otherServiceEntity->setServiceNo($serviceNumber['serviceNo']);
-            } else {
-                // create
-                $otherServiceEntity = new BusRegOtherService();
-                $otherServiceEntity->setBusReg($busReg);
-                $otherServiceEntity->setServiceNo($serviceNumber['serviceNo']);
-            }
-
-            $this->getRepo('BusRegOtherService')->save($otherServiceEntity);
-            $reduced[] = $otherServiceEntity->getId();
-        }
-
-        // remove the remaining records
-        foreach ($busReg->getOtherServices() as $otherServiceEntity) {
-            if (!in_array($otherServiceEntity->getId(), $reduced)) {
-                $this->getRepo('BusRegOtherService')->delete($otherServiceEntity);
+            // remove the remaining records
+            foreach ($busReg->getOtherServices() as $otherServiceEntity) {
+                if (!in_array($otherServiceEntity->getId(), $reduced)) {
+                    $this->getRepo('BusRegOtherService')->delete($otherServiceEntity);
+                }
             }
         }
